+++ src/directives/index.ts
@@ -12,6 +12,7 @@ export * from './headerContainerTemplate/headerContainerTemplate.directive';
 export * from './headerRowContainerTemplate/headerRowContainerTemplate.directive';
 export * from './matrixGridColumn/matrixGridColumn.directive';
 export * from './orderable/orderable.directive';
+export * from './showMetadataSelectorFor/showMetadataSelectorFor.directive';
 export * from './tableGridBodyCellTemplate/tableGridBodyCellTemplate.directive';
 export * from './tableGridHeaderCellTemplate/tableGridHeaderCellTemplate.directive';
 export * from './useTable/useTable.directive';